fix: Refactor as per PR comments.
