+++ src/components/shared/ContractFunction.vue
@@ -1,9 +1,8 @@
 <script setup lang="ts">
 import useWeb3 from '@/compositions/useWeb3'
-import { BN } from 'bn.js'
+import { decodeData, MethodType } from '@/helpers/functionUtils'
 import { reactive, computed, watch, onMounted } from 'vue'
-import { toWei, Unit } from 'web3-utils'
-import { MethodSelect, MethodType } from '@/endpoints/SendTransaction.vue'
+import { toWei, Unit, padLeft } from 'web3-utils'
 import ParamField from './ParamField.vue'
 
 interface ElementType extends MethodType {
@@ -31,109 +30,6 @@ const emits = defineEmits<Emits>()
 
 const { getWeb3 } = useWeb3()
 const { eth } = getWeb3()
-
-const SIGNATURE_CACHE = 'signature-cache'
-
-type BytesSignatureResponse = {
-  count: number
-  next: unknown
-  previous: unknown
-  results: [
-    {
-      id: number
-      created_at: string
-      text_signature: string
-      hex_signature: string
-      bytes_signature: string
-    }
-  ]
-}
-
-const fetcher = async <Response, Request>(config: {
-  url: string
-  method: 'GET' | 'POST'
-  data?: Request
-  headers?: Record<string, never>
-}): Promise<Response> => {
-  const fetchConfig: RequestInit = {
-    method: config.method,
-    headers: config.headers || {
-      Accept: 'application/json',
-      'Content-Type': 'application/json',
-    },
-  }
-  if (config.data) {
-    fetchConfig.body = JSON.stringify(config.data)
-  }
-
-  const response = await fetch(config.url, fetchConfig)
-
-  if (!response.ok) {
-    return response
-      .json()
-      .catch(() => {
-        throw new Error(response.status.toString())
-      })
-      .then(message => {
-        throw message
-      })
-  }
-  return await response.json()
-}
-
-const decodeData = async (data: string): Promise<MethodSelect> => {
-  if (/^0x/i.test(data)) {
-    data = data.substring(2)
-  }
-  const selector = data?.substring(0, 8)
-  if (selector) {
-    const signatureCache = await caches.open(SIGNATURE_CACHE)
-    const url = `https://www.4byte.directory/api/v1/signatures/?hex_signature=${selector}`
-    const functionSignatureResponse = await signatureCache.match(url)
-    if (functionSignatureResponse) {
-      return await functionSignatureResponse.json()
-    }
-
-    const methods = await fetcher<BytesSignatureResponse, void>({
-      method: 'GET',
-      url,
-    })
-
-    if (methods && methods.results.length > 0) {
-      for (const result of methods.results) {
-        try {
-          const params: string[] = result.text_signature
-            .replace(/^[^(]*\(|\)[^)]*$/g, '')
-            .split(',')
-          const args = eth.abi.decodeParameters(params, data.substring(8))
-          const encodeArgs = Array(params.length)
-            .fill(null)
-            .map((_val, index) => args[`${index}`] ?? '0x')
-          const newData = eth.abi
-            .encodeParameters(params, encodeArgs)
-            .substring(2)
-          if (data.substring(8) === newData) {
-            const item = {
-              label: `Decoded ${result.text_signature.replace(/\(.*$/, '')}`,
-              call: result.text_signature.replace(/\(.*$/, ''),
-              inputs: params.map((type, index) => ({
-                type,
-                name: `arg${index + 1}`,
-                value: args[index],
-              })),
-            }
-            await signatureCache.put(url, new Response(JSON.stringify(item)))
-            return item
-          }
-        } catch (err) {
-          // Ignore to try next record
-          console.error(err)
-        }
-      }
-    }
-  }
-  throw new Error(`Unable to decode data`)
-}
 
 function convertModel(model?: MethodType[]) {
   model =
@@ -256,7 +152,7 @@ function handleCall(e: Event) {
 const makeBytes32 = (value: string, type: string) => {
   if (type === 'bytes32') {
     if (/^[0-9]*$/.test(value)) {
-      return `0x${new BN(value).toString('hex', 64)}`
+      return padLeft(value, 64)
     }
   }
   return value
@@ -312,7 +208,7 @@ watch(
     const { value: _value } = output.value
     if (value && value !== _value) {
       try {
-        const method = await decodeData(value)
+        const method = await decodeData(eth, value)
         data.call = method.call
         data.items = method.inputs || []
       } catch (err) {
