+++ src/helpers/constants.ts
@@ -1,0 +1,19 @@
+export const sampleEoA = '0x311611C9A46a192C14Ea993159a0498EDE5578aC'
+export const sampleUP = '0xe608aBEeB2EA0EBb59170de6CBcFFaE06437fE0c'
+export const sampleSC = '0xcAC51571007DaAB53f26C2387b3B16420491dE18'
+// ERC20
+export const erc20TokenWithEip165 = '0xF5443372766a48faF098244c8C769c5AEa02f321'
+export const erc20TokenWithoutEip165 =
+  '0xB29c50a9F3D90FA3aDF394f2960BD6D8e0Ff5E9D'
+// ERC777
+export const erc777TokenWithEip165 =
+  '0xC719f454C8F9a0C7eEC4203B21766B88d8a5B073'
+export const erc777TokenWithoutEip165 =
+  '0xD7549C70A6122cA01043831f0f0c65152C4877d6'
+// LSP7
+export const lsp7TokenDivisible = '0x314E7a56B08AF8E729612930dBAd70BB5A3575D9'
+export const lsp7TokenNonDivisible =
+  '0xF5d8FD6599Cb1971b8EEba218FFE31da34a257a9'
+// ERC721
+export const erc721TokenWithEip165 =
+  '0x57b8e4f3C96180088652dc361473bB91266bb080'
